chore: code cleanup, introduce MultiIssuerJwtAuthenticationConverter
